use big int for center code

+++ GaelO2/database/migrations/2020_07_30_134121_create_center_user_table.php
@@ -16,7 +16,7 @@ class CreateCenterUserTable extends Migration
         Schema::create('center_user', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id')->nullable(false);
-            $table->unsignedInteger('center_code')->nullable(false);
+            $table->unsignedBigInteger('center_code')->nullable(false);
             $table->timestamps();
             //Dependencies
             $table->foreign('center_code')->references('code')->on('centers');
